extract common utils from batchSyncer
