+++ pkg/neg/syncers/batch_test.go
@@ -162,140 +162,6 @@ func TestToZoneNetworkEndpointMap(t *testing.T) {
 	}
 }
 
-func TestEncodeDecodeEndpoint(t *testing.T) {
-	ip := "10.0.0.10"
-	instance := "somehost"
-	port := "8080"
-
-	retIp, retInstance, retPort := decodeEndpoint(encodeEndpoint(ip, instance, port))
-
-	if ip != retIp || instance != retInstance || retPort != port {
-		t.Fatalf("Encode and decode endpoint failed. Expect %q, %q, %q but got %q, %q, %q.", ip, instance, port, retIp, retInstance, retPort)
-	}
-}
-
-func TestCalculateDifference(t *testing.T) {
-	testCases := []struct {
-		targetSet  map[string]sets.String
-		currentSet map[string]sets.String
-		addSet     map[string]sets.String
-		removeSet  map[string]sets.String
-	}{
-		// unchanged
-		{
-			targetSet: map[string]sets.String{
-				negtypes.TestZone1: sets.NewString("a", "b", "c"),
-			},
-			currentSet: map[string]sets.String{
-				negtypes.TestZone1: sets.NewString("a", "b", "c"),
-			},
-			addSet:    map[string]sets.String{},
-			removeSet: map[string]sets.String{},
-		},
-		// unchanged
-		{
-			targetSet:  map[string]sets.String{},
-			currentSet: map[string]sets.String{},
-			addSet:     map[string]sets.String{},
-			removeSet:  map[string]sets.String{},
-		},
-		// add in one zone
-		{
-			targetSet: map[string]sets.String{
-				negtypes.TestZone1: sets.NewString("a", "b", "c"),
-			},
-			currentSet: map[string]sets.String{},
-			addSet: map[string]sets.String{
-				negtypes.TestZone1: sets.NewString("a", "b", "c"),
-			},
-			removeSet: map[string]sets.String{},
-		},
-		// add in 2 zones
-		{
-			targetSet: map[string]sets.String{
-				negtypes.TestZone1: sets.NewString("a", "b", "c"),
-				negtypes.TestZone2: sets.NewString("e", "f", "g"),
-			},
-			currentSet: map[string]sets.String{},
-			addSet: map[string]sets.String{
-				negtypes.TestZone1: sets.NewString("a", "b", "c"),
-				negtypes.TestZone2: sets.NewString("e", "f", "g"),
-			},
-			removeSet: map[string]sets.String{},
-		},
-		// remove in one zone
-		{
-			targetSet: map[string]sets.String{},
-			currentSet: map[string]sets.String{
-				negtypes.TestZone1: sets.NewString("a", "b", "c"),
-			},
-			addSet: map[string]sets.String{},
-			removeSet: map[string]sets.String{
-				negtypes.TestZone1: sets.NewString("a", "b", "c"),
-			},
-		},
-		// remove in 2 zones
-		{
-			targetSet: map[string]sets.String{},
-			currentSet: map[string]sets.String{
-				negtypes.TestZone1: sets.NewString("a", "b", "c"),
-				negtypes.TestZone2: sets.NewString("e", "f", "g"),
-			},
-			addSet: map[string]sets.String{},
-			removeSet: map[string]sets.String{
-				negtypes.TestZone1: sets.NewString("a", "b", "c"),
-				negtypes.TestZone2: sets.NewString("e", "f", "g"),
-			},
-		},
-		// add and delete in one zone
-		{
-			targetSet: map[string]sets.String{
-				negtypes.TestZone1: sets.NewString("a", "b", "c"),
-			},
-			currentSet: map[string]sets.String{
-				negtypes.TestZone1: sets.NewString("b", "c", "d"),
-			},
-			addSet: map[string]sets.String{
-				negtypes.TestZone1: sets.NewString("a"),
-			},
-			removeSet: map[string]sets.String{
-				negtypes.TestZone1: sets.NewString("d"),
-			},
-		},
-		// add and delete in 2 zones
-		{
-			targetSet: map[string]sets.String{
-				negtypes.TestZone1: sets.NewString("a", "b", "c"),
-				negtypes.TestZone2: sets.NewString("a", "b", "c"),
-			},
-			currentSet: map[string]sets.String{
-				negtypes.TestZone1: sets.NewString("b", "c", "d"),
-				negtypes.TestZone2: sets.NewString("b", "c", "d"),
-			},
-			addSet: map[string]sets.String{
-				negtypes.TestZone1: sets.NewString("a"),
-				negtypes.TestZone2: sets.NewString("a"),
-			},
-			removeSet: map[string]sets.String{
-				negtypes.TestZone1: sets.NewString("d"),
-				negtypes.TestZone2: sets.NewString("d"),
-			},
-		},
-	}
-
-	for _, tc := range testCases {
-		addSet, removeSet := calculateDifference(tc.targetSet, tc.currentSet)
-
-		if !reflect.DeepEqual(addSet, tc.addSet) {
-			t.Errorf("Failed to calculate difference for add, expecting %v, but got %v", tc.addSet, addSet)
-		}
-
-		if !reflect.DeepEqual(removeSet, tc.removeSet) {
-			t.Errorf("Failed to calculate difference for remove, expecting %v, but got %v", tc.removeSet, removeSet)
-		}
-	}
-}
-
 func TestSyncNetworkEndpoints(t *testing.T) {
 	syncer := NewTestSyncer()
 	if err := syncer.ensureNetworkEndpointGroups(); err != nil {
